+++ src/compas/geometry/primitives/plane.py
@@ -1,11 +1,13 @@
-from __future__ import print_function
 from __future__ import absolute_import
 from __future__ import division
+from __future__ import print_function
 
 from math import sqrt
+
+from compas.geometry import cross_vectors
+from compas.geometry.primitives import Point
 from compas.geometry.primitives import Primitive
 from compas.geometry.primitives import Vector
-from compas.geometry.primitives import Point
 
 
 class Plane(Primitive):
@@ -204,7 +206,7 @@ class Plane(Primitive):
         a = Point(*a)
         b = Point(*b)
         c = Point(*c)
-        normal = Vector.cross(b - a, c - a)
+        normal = Vector(*cross_vectors(b - a, c - a))
         return cls(a, normal)
 
     @classmethod
@@ -228,14 +230,14 @@ class Plane(Primitive):
 
         Examples
         --------
-        >>> plane = Plane.from_three_points([0.0, 0.0, 0.0], [2.0, 1.0, 0.0], [0.0, 3.0, 0.0])
+        >>> plane = Plane.from_point_and_two_vectors([0.0, 0.0, 0.0], [1.0, 0.0, 0.0], [0.0, 1.0, 0.0])
         >>> plane.point
         Point(0.000, 0.000, 0.000)
         >>> plane.normal
         Vector(0.000, 0.000, 1.000)
 
         """
-        normal = Vector.cross(u, v)
+        normal = Vector(*cross_vectors(u, v))
         return cls(point, normal)
 
     def from_abcd(cls, abcd):